OK; so look and walk is "implemented", but I don't really like the feel of it.
It could use some work - as in perhaps we should walk as per usual (keyboard) and then
look if we aim (right key) in the direction of look.

And I do not like the animation state machine. Need something else. Perhaps
stack of states or more stopevents, obviously, more stop events!

Or?